noop

+++ js/list.js
@@ -30,15 +30,17 @@ angular.module('list', ['kings-app.utils'])
     }
 
     function deleteData(data){
-      builtApi.deleteObject({
-        options :{
-          classUid : classUid,
-          objectsUid: data.uid
-        }
-      }).then(function(res){
-        console.log("res");
-        $scope.newLists.splice($scope.newLists.indexOf(data), 1);
-      })
+      if(confirm("Are you sure you want to delete this object?")){
+        builtApi.deleteObject({
+          options :{
+            classUid : classUid,
+            objectsUid: data.uid
+          }
+        }).then(function(res){
+          console.log("res");
+          $scope.newLists.splice($scope.newLists.indexOf(data), 1);
+        })
+      }
       
     }
